Removed License Section
I missed this entirely, but according to the official pull request guidelines. "Do not add the license name, text, or a License section to the readme. GitHub already shows the license name and link to the full text at the top of the repo." The license section has been removed.