Exclude Itch.io in link check workflow (for now)
Unfortunately, I just don't see a way around this. The workflow should be fine for every other site except for itch. Itch links will need to be checked manually.
Inside the workflow job, excluded links are marked with [EXCLUDED]. In the future, if there are new tools to deal with too many request errors, or itch's rate limiting becomes less aggressive, we can revisit this.